Implement IP binding by adding the HTTP_IP variable
By default, the exporter will bind to all IP (::).
Add the HTTP_IP variable to be able to listen on a specific IP.

If the HTTP_IP variable does not contain a valid IP address, it will
be ignored and the daemon will listen on ::

This variable accept IPv4 and IPv6 address.